requirements.txt
requests
beautifulsoup4
lxml
python-dotenv
pymysql
python-dateutil